Adding a first version of JNAL. Only one function (jnal::Play) which can play OGG files.
git-svn-id: file:///private-backup/var/svn/jngl@94 8b392cba-0c28-4fbc-ad53-5d9c93102ef3

+++ freetype.cpp
@@ -289,7 +289,7 @@ namespace jngl
 
 	void Font::Print(const int x, const int y, const std::string& text)
 	{
-		double h = height_ / .63; // We make the height about 1.5* that of
+		int h = int(height_ / .63); // We make the height about 1.5* that of
 
 		std::vector<std::string> lines(ParseString(text));
